HDS-134-Reorder Option shows inconsistent qty added to cart
https://four51.atlassian.net/browse/HDS-134
This was happening when the reorder function is hit, the pop up is not updating line item quantities as they are being called.  This solution aggregates the line items that are being added in the first half second and then displays the aggregate quantity amount and leaves room to tweak the timing if needed.