HDS-324-
I think the issue is that query params are no longer on the url when routing to the product page, thereby not triggering the function which updates the breadcrumbs - since its not updating- your seeing the previous catergory a user was on.

fixed by checking on initial component load if query params are empty- if so, clear breadcrumbs. since if there are query params the bread crumbs will update correctly

https://four51.atlassian.net/browse/HDS-324